Move NetworksTestCase to test_networks and add tests for network admins.
